LU-17245 utils: fix lfs error messages with multiple paths
When using some lfs utilities (find, getstripe) with multiple paths,
if a subset of the paths has an error, for example due to a typo, the
error message produced can be misleading. For "getstripe" it refers to
the last file on the command line regardless of which file had the
error, and for "find" it prints out the right filename but uses
the error code from the last file on the command line.

This cleans up these error messages for "lfs find" and
"lfs getstripe".

This also adjusts "lfs setdirstripe" to continue for subsequent files
if it encounters an error for earlier files on the command line.

Signed-off-by: Thomas Bertschinger <bertschinger@lanl.gov>
Fixes: bc500536b6dd ("LU-930 utils: fix 'lfs find' error message")
Fixes: 4affa48f676b ("LU-5170 utils: Continue on error when multiple files requested")
Fixes: a24f61532927 ("LU-11213 dne: add new dir hash type "space"")
Change-Id: I9cdd007912ffb4f6ebc31e422851977e49186ae7
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/52942
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>