LU-17263 utils: 'lfs find -size' to use 512-byte units
Change the 'lfs find -size' argument to 512-byte blocks by default if
no unit is given.  This better matches find(1) and avoids confusion
when converting "find" arguments to "lfs find".  Accept the 'c' suffix
like find(1) to specify a number of characters (bytes).

Most users/scripts will specify a unit, so it is expected not to cause
significant upset with this change.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I3124e667acc06928f41a3d3006e1d9b4a43ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/52994
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Anjus George <georgea@ornl.gov>
Reviewed-by: Oleg Drokin <green@whamcloud.com>