LU-16967 build: Add in-kernel-ko2iblnd driver
Add in-kernel-ko2iblnd.ko for users of in-kernel OFED
and only build ko2iblnd.ko if an external OFED is available

This allows for building and packaging both an external
(MOFED or OPA) o2ib driver and an in-kernel o2ib driver.

Packaging rules will be written so that only enable one
of the o2iblnd drivers can be installed.

In the case of the in-kernel-ko2iblnd.ko driver a symlink
named ko2iblnd.ko will be created to point to the in-kernel
based o2ib driver which allows for a reasonable migration path
for the majority of users.

It is useful for dist build and test to be able to build
both in-kernel IB and external OFED in the same build.

This also means there would be some install/configure
adjustments that ought to have some discussion.

Test-Parameters: trivial
HPE-bug-id: LUS-11711
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I8105fad0b20c36705d7e14e3ae976bf3d81e9f1b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/51915
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>