LU-13138 tests: improve sanity/101d 'dd' parsing
If 'dd' takes a long time to complete, or if it finishes in an
exact number of seconds, it will not print a decimal point, so
the current regexp will fail to detect the runtime.

Improve test_101d to allow parsing the 'dd' runtime in this case.

Test-Parameters: trivial testlist=sanity env=ONLY=101d,ONLY_REPEAT=100
Fixes: 43ebfad490 ("LU-13138 tests: measure 'dd' time more accurately")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Icb030467c76947d0546916e11a91e5afb33ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53188
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Alex Deiter <alex.deiter@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>