LU-16739 uapi: restore LUSTRE_NODEMAP_NAME_LENGTH in lustre_disk.h
sanity test 400b fails with:

lustre_disk.h:266:18: error: 'LUSTRE_NODEMAP_NAME_LENGTH' undeclared here (not in a function)
 char   ncr_name[LUSTRE_NODEMAP_NAME_LENGTH + 1];

This is due to the move of LUSTRE_NODEMAP_NAME_LENGTH to lustre_idl.h.
Move it back and this time add a message to avoid this in the
future.

Test-Parameters: trivial
Fixes: 8d828762d1 ("LU-17015 gss: support large kerberos token for rpc sec init")
Change-Id: I5479bbf13f26bfd3b4f6e5f7c0c1688d810fca53
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53208
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>